feat(android): 홈 화면 구현

+++ shared/src/commonMain/kotlin/good/space/runnershi/ui/components/QuestCard.kt
@@ -30,7 +30,7 @@ import org.jetbrains.compose.ui.tooling.preview.Preview
 @Composable
 fun QuestCard(
     title: String,
-    exp: Int,
+    exp: Long,
     isCleared: Boolean = false,
     description: String? = null,
     modifier: Modifier = Modifier
